Buggy test files folder, cleanup (use ./compile, ./valgrind, ./diff)

+++ wet1/code/Movie.h
@@ -44,7 +44,7 @@ private:
     bool m_vipOnly;
     int m_id;
     double m_ratingAverage;
-    double m_ratingCount;
+    int m_ratingCount;
     Genre m_genre;
 };
 #endif //__MOVIE_H__
